Rename "TimeBarChart" to "TimeChart"
棒グラフと折れ線グラフをどちらも含むコンポーネントになったので，BarChartという名前が不適切になった
時系列データという意味でTimeChartにした
